UserRegFlow - Enhanced Form State and Validation
(1) Form State. The from in DonorForm holds the main state. This is
provided to the sub-forms as default values. In return, the sub-forms
always call onSave to update the main form with their state. With this
arrangement, previously entered values can be restored when a user
navigates between the sub-forms; thereby achieving the illusion of
persisted sub-form states.

(2) Form Validation. Validation occurs in the sub-forms when
progressing forwards (i.e. Next or Submit). When progressing forward,
validation occurs before onSave. Thus, the main form will see only
validated values.