Improved Vet Options for Registration Form
(1) Feature - Vet addresses are now rendered below the vet name. The
BarkFormOption type was extended to include an optional `description`
to support this. When `description` is specified, a FormDescription is
rendered below the FormLabel. Currently this only works for radio
groups.

(2) Testing - The function for querying the database is moved into
./_lib/get-vet-options so its getVetOptions function can be tested.

(3) Refactoring - The manner by which the async retrievals for breed
list and vet options are orchestrated is improved.