Fix icon scaling in my account page (#112)
The icons are of different shapes. Yet when we set the dimensions to the
exact values from the SVG files, the alignment with the text alongside
on the right goes haywire. Thus we put the icons in a 25x25 div and
apply full/auto on width/height as appropriate. Now, all of these
styling is very verbose. So the icons and values are put into a
`userDetails` array and rendered consistently using `map()`.