Update sidebar layout to no longer add padding around main content (#117)
In view of the design for /user/info, where there is a section with
coloured background that's flush to the left and right sides of the
page, this commit removes the padding that the sidebar layout used to
place around the main content. Each page should not be responsible for
adding their own padding or margins.