Extend Call Outcome concept to include REPORTED and CANCELLED posibilities (#121)
(1) Extended the concept of call-outcome to include the REPORTED and
CANCELLED to call outcomes. Now APPOINTMENT means there a report or
cancellation is pending. Thus, when call-outcome is APPOINTMENT, there's
no need to check for a report.

(2) Consequently, when a report is inserted, the call-outcome must be
updated to REPORTED. To facilitate this, the dbInsertReport function has
been replaced with dbInsertReportAndUpdateCall.

(3) Additionally, for the purpose of retrieving or counting pending
reports, it is only necessary to check if a call's outcome is
APPOINTMENT. No longer is it necessary to check if a call's ID is absent
from the reports table. Thus several queries that used to do this have
been updated in this commit to stop doing that.