Resolve latest weight correctly (#126)
(1) This commit fixes the way latest weight is resolved.

(1.1) Correct way is to use the latest value from EITHER the "dogs"
table or from the "reports" table.

(1.2) The wrong way, which this fixes was to prefer the latest value
from the "reports" table.

(2) To fix this, a new "profile_modification_time" column had to be
added to the "dogs" table. This column defaults to CURRENT_TIMESTAMP. So
it is set automatically upon insertion. Subsequently, it is updated to
CURRENT_TIMESTAMP by updateDogProfile and updateSubProfile. The column
is then used in the "latest_values" view to resolve the correct
"latest_dog_weight_kg". (Related to barkbank-schema migrations V32 and
V33.)