Improve accessibility so iOS Safari does not zoom in on text inputs (#132)
(1) When the font size of a text input is under 16px, iOS Safari will
zoom into the text input when it is focused. This breaks the mobile
webapp experience because when the user is done entering their
information, they have to manually zoom out.

(2) This commit solves the problem by using "text-base" Tailwind CSS
class to set the font size to 16px. This is done for BarkFormInput and
BarkFormDateInput.