Show a dock instead of a sidebar when viewing on mobile (#131)
(1) Show a dock instead of a sidebar. Switch to sidebar for screens md
and above.

(2) A component called BarkNavLayout replaces the BarkSidebarLayout. The
ID associated with this is `#bark-nav`. It replaces `#bark-sidebar`.

(3) The SidebarComponent in the e2e tests is renamed to NavComponent. It
is updated to look for `#bark-nav`.

(4) Some padding is added to the bottom of the footer. This padding is
the height of the dock, but becomes 0 for screens md and larger. The
amount of padding and the height of the dock is something that needs to
be kept the same. (A little hacky, but an easy solution.)