Use toasts to indicate status of saving/adding dog (#146)
(1) Sometimes it can take awhile for data to get into the database.
(Another issue!) So, in the frontend we want to indicate when data is
"Saving..." and "Saved". In this commit, we use Toasts to accomplish
this.

(2) This is experimental. Therefore, only the add-dog and edit-dog save
sequences have been instrumented to use toasts.

(3) Variants have been defined for "brandInfo", "brandSuccess", and
"brandError". Not much thought has been put into the styling. This can
be revisited in time.