bark-button improvements (#152)
(1) Some changes to BarkButton

(1.1) The `onClick` callback type now allows both asynchronous and
synchronous functions.

(1.2) Added support for `disabled` buttons.

(1.3) Adjust the styling for `active:` and `hover:` states.

-(a) The earlier behaviour of switching between brand and brandInverse
styles is now use to distinguish the `active:` state.

-(b) The `hover:` state is now a light opacity change.

(1.4) `md:w-[160px]` is removed because this makes the button size fixed
for `md` screens and bigger—which became an issue when rendering buttons
in small spaces on big screens. In general, I think the size of a button
is related to the layout of the page and not the style of the button.