Implement vet appointment scheduler (#151)
(1) About — This commit implements a basic appointment scheduler for Vet
users. The scheduler supports the following functionality: (i) Vet can
list dogs that are available for scheduling; (ii) Vet can retrieve owner
contact details of available dogs; and (iii) Vet can record Scheduled or
Declined call outcomes.

(2) Details

(2.1) shadcn-ui — The following shadcn-ui components were installed: (i)
skeleton; (ii) scroll-area; and (iii) badge.

(2.2) BarkUserContactDetails — The elements presenting a user's contact
details in the View My Account page were extracted into a reusable
BarkUserContactDetails component so that it can be used to display user
contact details in the Vet Appointment Scheduler.

(2.3) Vet Navigation — This was updated for appointment scheduler and
also for add-reports later on. Here we deviated from the Figma design by
using the labels "Schedule" and "Report". This is because space in the
dock is limited.

(2.4) `DEV_DB_DELAY_MILLIS` — A new environment variable is supported to
introduce delays after DB queries. This is optional and defaults to 0.
The feature is useful for testing page or content loading states.

(2.5) recordCallOutcome — This vet actor action is for recording call
outcomes: APPOINTMENT | DECLINED.

(2.6) AppointmentScheduler — This is the component that shows the list
of available dogs and owner details of the selected dog in the list, if
any. It is fully a client component. I reckon that is needed for the
interactivity: (i) Highlighting the selected dog card; (ii) Showing
"Scheduled" / "Declined" badges, etc.