Impl bark domain abstraction layer (#158)
(1) A new bark abstraction layer is introduced to define the entities
and operations pertaining to the bark bank domain. The vision is to
eventually migrate all operations into this layer. Although it remains
to be seen whether that is the best course of action.

(2) Bark Abstraction Layer — This is located in the lib/bark module. At
the centre of this are the bark-context and bark-models.

(2.1) bark-context — Context is like the "this" or the "self". It
provides context for the functions located in "operations". We could
have gone with a BarkService (in fact we did), but when it is full of
methods it is expected to grow to thousands of lines.

(2.2) bark-models — This contains models. To separate them from the
rest, all the models in here have the "Bark" prefix. E.g. "BarkReport".

(3) Operations for creating, retrieving, and editing reports were
implemented.