Exploratory implementation of /vet/reports (#160)
(1) In this commit we introduce an exploratory implementation of a page
for vets to manage reports.

(2) Changes

(2.1) Path is changed from /vet/report to /vet/reports. The singular
form was intended to be a verb, the plural form is a noun. It was felt
that the noun form is a better fit for the idea of "report management".

(2.2) shadcn-ui Card was installed. Although, it remains to be seen how
useful it actually is.

(2.3) ReportsExplorer. The bulk of the change is in this component. It
is a client component. This is used to explore how to implement a list
of cards that become a "table" responsively. This is the extent of the
exploration really. The intention is to show it is possible so that we
can consider this option when designing the Reports tool (that for
managing reports).