Implement page for vets to edit reports (#176)
(1) Implement the page for vets to edit reports.

(2) Changes

(2.1) Collapse all the different `_components` and `_actions`
directories in `app/vet` into sub directories in `app/vet/_lib`. This
makes it easier to reuse components and code across `app/vet` frontend.

(2.2) Extract a GeneralReportForm out of SubmitReportForm. This is used
to implement the EditReportForm underlying the
`/vet/reports/edit/[reportId]` page.