Implement report view for vets (#177)
(1) Implemented the page for viewing a report. This is for vets.

(2) Other Changes

(2.1) The link to edit a report in the Report List page is removed.
Instead, clicking on a card brings the Vet user to the Report View page.
The Report View page has Back and Edit buttons. The Edit button will
bring the Vet to the Report Edit page. The UI test was updated to go
through this process.

(2.2) Defined a set of x-card classes in global CSS.