Do not show user email to vets (#178)
(1) User email is not useful for a call-list and it is also a login
credential. Thus, in this commit, user emails are removed from the vet
interfaces.

(2) Changes

(2.1) The BarkUserContactDetails component is updated to optionally
accept user email. Email is provided in My Account, but not when a vet
is viewing Owner Contact Details.

(2.2) The API for fetching owner contact details no longer provides user
email.