Revise the general report form (#182)
(1) This revises the form inputs in the General Report Form for Reported
Ineligibility. Previously, the form framed the questions as one about
Eligibility. This misses the point of Reported Ineligibility, which is
to ask for additional reasons why a dog might be ineligible.

(2) This also include the following additional improvements.

(2.1) Improved validation for Reported Ineligibility. —(a) Ask for
reason, then status, and then expiry date. —(b) Smartly check for reason
before using status. E.g. if the reason is empty and the status is
temporarily ineligible, the status recorded should still be NIL.

(2.2) Visit Time is changed to Visit Date.

(2.3) Improved cross-field validations. E.g. ineligibility expiry date
is required when temporarily ineligible, and only when temporarily
ineligible. Another example is checking that visit date is not in the
future. The full list can be found in the `schemaWithRefinements`
object.

(2.4) Improved presentation. The form is now split into three sections:
(i) Visit & Outcomes; (ii) Observations; and (iii) Ineligibility.