Preparations before adding report lists to the User View Dog page (#203)
(1) In order to add reports to the User View Dog page, we did some
refactoring and fixed a bug. This commit includes just the refactoring
and the bug fix. We'll work on adding reports after.

(2) Refactoring

- All db-enums have been migrated into bark/enums.
- All dog-models have been migrated into bark/models.
- The Zod schemas from the above were used to validate SQL queries in
existing user-actions.

(3) Bug fix: Fixed parsing of dog weight in General Dog Form. Previously
it was using parseFloat, which returns NaN values when given NULL. This
was discovered when Zod schema validation failed!

(4) There were also some UI changes to the dog view page: (i) Edit
button is replaced with an Edit Icon; and (ii) Back button is now "Back
to Pets" and placed outside of the Profile card. This will make it
easier to add reports later; we'll see.