Implement page for users to view reports (#207)
Implement a page for users to view dog medical reports.

(1) The approach taken was to first generalise the Vet's ReportView and
then reuse that for the User pages. This was achieved as follows:

-(a) The introduction and controls from ReportView were stripped out and
moved into the parent page.tsx file.

-(b) The ReportView was moved to app/_components; i.e. a location that
is neither Vet nor User specific.

-(c) The User's view-report page.tsx component was updated to use the
now neutral ReportView.

(2) In addition to the visual changes, the opFetchReport function was
extended to accept an actorUserId value. This is used to validate
ownership of the dog for which the report was written. When ownership
fails, an ERROR_WRONG_OWNER error would be returned.

(3) Some refactoring to the UI test code was done. Moving and renaming
models like GeneratedDog to a models directory. This work was left half
done because it deserves its own task, its own commit. We will revisit
this when we work on writing the UI test for "User can view report".