Add support for editing sub profiles (#209)
When a dog has a report, users are allowed to edit only some of the
profile fields. These are known as sub profiles. In this commit, the
edit-dog page is updated to consider the existing report count, and when
this is non-zero, it will present to the user a form for editing a
sub-profile.