Fix BarkButton and unstable UI tests (#212)
(1) BarkButton changes to addressing the following issues:

-(a) Saving form changes instead of cancelling (vet edit report page).

-(b) Login form validates OTP field when user clicks "Send me an OTP".

(1.1) Update BarkButton props to be either for the link kind or the
button kind. For the link kind, 'href' is specified and 'type' and
'onClick' can be omitted or must be set to undefined. For the button
kind, 'type' must be specified, 'href' can be omitted or must be
undefined, and 'onClick' is optional.

(1.2) Update all usage that isn't a link—i.e. no 'href'—to specify
'type' as either 'button' or 'submit'. Next Build will fail if we do not
provide the correct props.

(2) Fixed some flaky UI tests. Specifically, the "user can view report
and edit sub-profile" test. Here, the "doSubmitReport" sequence was not
checking the URL after clicking the submit button. Thus the page
transitions raced with the expectations of the follow-up
doLogoutSequence. Changes here include the following:

-(a) Updates to doLogoutSequence to use doGetIsMobile

-(b) Add checkUrl in at the end of some of the UI test ops.