HKDF Encryption Protocol with support for key-rotation (#218)
It is more efficient to use HKDF to derive keys than it is to use
PBKDF—which is slow by design. In this commit, HKDF is introduced as the
preferred key derivation function.

Changes

(1) A new lower level abstraction called EncryptionProtocol is
introduced that is "used by" EncryptionService. The previous PBKDF
implementation of EncryptionService—SecretEncryptionService—was then
ported into a PbkdfEncryptionProtocol and ultimately removed.

(2) A MultiProtocolEncryptionService is introduced that holds a chain of
EncryptionProtocol. This provides the avenue through which we may
introduce new protocols.

(3) A HkdfEncryptionProtocol is implemented.