BARK-5778: Improve UI Test Stability (#228)
(1) Recently UI tests have been flaky we addressed this with two
different approaches.

(i) We increased the timeout of some slow tests to 1 minutes.

(ii) We introduced a way for pages to define how to checkPageLoaded.
Since approach (i) resolved the issues, we only used checkPagePayload
for a few pages. Nevertheless, the feature is available for future
tests.